refactor: argument handling no uses modules
Improved readability
Cleaner main.rs
