+++ src/lib.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+extern crate structopt;
+
+// -- Personel mods
+pub mod model;
